fix(logs): Redacted secret data in logs.
