+++ fiat-api/src/main/java/com/netflix/spinnaker/fiat/shared/FiatAccessDeniedExceptionHandler.java
@@ -19,8 +19,6 @@ package com.netflix.spinnaker.fiat.shared;
 import com.netflix.spinnaker.kork.api.exceptions.AccessDeniedDetails;
 import com.netflix.spinnaker.kork.web.exceptions.ExceptionMessageDecorator;
 import java.io.IOException;
-import java.util.Enumeration;
-import java.util.HashMap;
 import java.util.Map;
 import java.util.StringJoiner;
 import javax.servlet.http.HttpServletRequest;
@@ -38,12 +36,9 @@ public class FiatAccessDeniedExceptionHandler {
   private final Logger log = LoggerFactory.getLogger(getClass());
 
   private final DefaultErrorAttributes defaultErrorAttributes = new DefaultErrorAttributes();
+  private final HeadersRedactor headersRedactor = new HeadersRedactor();
 
   private final ExceptionMessageDecorator exceptionMessageDecorator;
-
-  private static final String SECRET_DATA_HEADER_FORMAT = "X-";
-  private static final String AUTHORIZATION_HEADER = "Authorization";
-  private static final String SECRET_DATA_VALUE = "**REDACTED**";
 
   public FiatAccessDeniedExceptionHandler(ExceptionMessageDecorator exceptionMessageDecorator) {
     this.exceptionMessageDecorator = exceptionMessageDecorator;
@@ -55,7 +50,7 @@ public class FiatAccessDeniedExceptionHandler {
       throws IOException {
     storeException(request, response, e);
 
-    Map<String, String> headers = requestHeaders(request);
+    Map<String, String> headers = headersRedactor.getRedactedHeaders(request);
 
     log.error(
         "Encountered exception while processing request {}:{} with headers={}",
@@ -106,32 +101,6 @@ public class FiatAccessDeniedExceptionHandler {
 
     if (authorizationFailure.hasAuthorization()) {
       sj.add("- required authorization:").add(authorizationFailure.getAuthorization().toString());
-    }
-  }
-
-  private Map<String, String> requestHeaders(HttpServletRequest request) {
-    Map<String, String> headers = new HashMap<>();
-
-    if (request.getHeaderNames() != null) {
-      for (Enumeration<String> h = request.getHeaderNames(); h.hasMoreElements(); ) {
-        String headerName = h.nextElement();
-        String headerValue = getRedactedHeaderValue(headerName, request.getHeader(headerName));
-        headers.put(headerName, headerValue);
-      }
-    }
-
-    return headers;
-  }
-
-  private String getRedactedHeaderValue(String headerName, String headerValue) {
-    if (headerName.startsWith(SECRET_DATA_HEADER_FORMAT)) {
-      return SECRET_DATA_VALUE;
-    } else {
-      if (headerName.contains(AUTHORIZATION_HEADER)) {
-        return headerValue.trim().split("\n")[0].split(" ")[0] + " " + SECRET_DATA_VALUE;
-      } else {
-        return headerValue;
-      }
     }
   }
